backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2 
django-filter==23.4
djoser==2.1.0
gunicorn==20.1.0
markdown==3.5.1
Pillow==10.1.0
psycopg2-binary==2.9.3
webcolors==1.11.1
drf-extra-fields==3.4.0
python-dotenv==1.0.0
django-colorfield==0.11.0